+++ aggregator/src/aggregator/http_handlers/tests/aggregation_job_init.rs
@@ -279,7 +279,7 @@ async fn aggregate_init_sync() {
     let (prepare_init_4, _) = prep_init_generator.next(&measurement);
 
     // prepare_init_5 falls into a batch that has already been collected.
-    let past_clock = MockClock::new((task.time_precision().as_seconds() / 2) as i64);
+    let past_clock = MockClock::new(task.time_precision().as_seconds() / 2);
     let report_metadata_5 = ReportMetadata::new(
         random(),
         past_clock
